feat(cmd): add a retry when ask for password
