+++ db/users/create.go
@@ -3,9 +3,6 @@ package users
 import (
 	"context"
 	"fmt"
-	"os"
-
-	"golang.org/x/term"
 
 	"github.com/Scalingo/cli/config"
 	"github.com/Scalingo/cli/io"
@@ -30,14 +27,9 @@ func CreateUser(ctx context.Context, app, addonUUID, username string, readonly b
 		return nil
 	}
 
-	password, confirmedPassword, err := askForPassword(ctx)
+	password, confirmedPassword, err := askForPasswordWithRetry(ctx, 3)
 	if err != nil {
-		return errors.Wrap(ctx, err, "ask for password")
-	}
-
-	passwordValidation, ok := isPasswordValid(password, confirmedPassword)
-	if !ok && password != "" {
-		io.Error(passwordValidation)
+		io.Error(err)
 		return nil
 	}
 
@@ -72,39 +64,4 @@ func CreateUser(ctx context.Context, app, addonUUID, username string, readonly b
 
 	fmt.Printf("User \"%s\" created.\n", databaseUsers.Name)
 	return nil
-}
-
-func askForPassword(ctx context.Context) (string, string, error) {
-	fmt.Printf("Password (Will be generated if left empty): ")
-
-	password, err := term.ReadPassword(int(os.Stdin.Fd()))
-	if err != nil {
-		return "", "", errors.Wrap(ctx, err, "read password")
-	}
-
-	fmt.Printf("\nPassword Confirmation: ")
-	confirmedPassword, err := term.ReadPassword(int(os.Stdin.Fd()))
-	if err != nil {
-		return "", "", errors.Wrap(ctx, err, "read password confirmation")
-	}
-	fmt.Println()
-
-	return string(password), string(confirmedPassword), nil
-}
-
-func isPasswordValid(password, confirmedPassword string) (string, bool) {
-	if password != confirmedPassword {
-		return "Password confirmation doesn't match", false
-	}
-	if len(password) < 8 || len(password) > 64 {
-		return "Password must contain between 8 and 64 characters", false
-	}
-	return "", true
-}
-
-func isUsernameValid(username string) (string, bool) {
-	if len(username) < 6 || len(username) > 32 {
-		return "name must contain between 6 and 32 characters", false
-	}
-	return "", true
 }
